backend/pyproject.toml
[tool.poetry]
name = "elon-back"
version = "0.1.0"
description = ""
authors = ["p_kozhevnikov <kozhevnikov0001@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.7"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"
ruff = "^0.5.2"
pytest = "^8.2.2"
pytest-django = "^4.8.0"
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
